+++ scqubits/core/param_sweep.py
@@ -20,11 +20,11 @@ import scqubits.core.descriptors as descriptors
 import scqubits.core.hilbert_space as hspace
 import scqubits.core.spec_lookup as spec_lookup
 import scqubits.core.storage as storage
-import scqubits.io.file_io_qutip
+import scqubits.io_utils.fileio_qutip
 import scqubits.settings as settings
 import scqubits.utils.cpu_switch as cpu_switch
-import scqubits.io.file_io as io
-import scqubits.io.file_io_serializers as serializers
+import scqubits.io_utils.fileio as io
+import scqubits.io_utils.fileio_serializers as serializers
 import scqubits.utils.misc as utils
 
 if settings.IN_IPYTHON:
@@ -365,7 +365,7 @@ class ParameterSweep(ParameterSweepBase, dispatch.DispatchClient, serializers.Se
             hamiltonian += self._hilbertspace.interactionterm_hamiltonian(interaction_term,
                                                                           evecs1=evecs1, evecs2=evecs2)
         evals, evecs = hamiltonian.eigenstates(eigvals=self.evals_count)
-        evecs = evecs.view(scqubits.io.file_io_qutip.QutipEigenstates)
+        evecs = evecs.view(scqubits.io_utils.fileio_qutip.QutipEigenstates)
         return evals, evecs
 
     def _lookup_bare_eigenstates(self, param_index, subsys, bare_specdata_list):
