+++ scqubits/io_utils/fileio.py
@@ -1,4 +1,4 @@
-# file_io.py
+# io_utils.py
 #
 # This file is part of scqubits.
 #
@@ -15,7 +15,7 @@ Helper routines for writing data to files.
 import os
 
 import scqubits.core.constants as const
-import scqubits.io.file_io_serializers as io_serializers
+import scqubits.io_utils.fileio_serializers as io_serializers
 
 
 def serialize(the_object):
@@ -123,7 +123,7 @@ class FileIOFactory:
         -------
         IOWriter
         """
-        import scqubits.io.file_io_backends as io_backends
+        import scqubits.io_utils.fileio_backends as io_backends
         _, suffix = os.path.splitext(file_name)
         if suffix == '.csv':
             return io_backends.CSVWriter(file_name)
@@ -149,7 +149,7 @@ class FileIOFactory:
         if get_external_reader:
             return get_external_reader(file_name, file_handle=file_handle)
 
-        import scqubits.io.file_io_backends as io_backends
+        import scqubits.io_utils.fileio_backends as io_backends
         _, suffix = os.path.splitext(file_name)
         if suffix == '.csv':
             return io_backends.CSVReader()
